+++ spock-core/src/main/java/org/spockframework/runtime/ExpressionInfoValueRenderer.java
@@ -14,16 +14,19 @@
 
 package org.spockframework.runtime;
 
-import org.spockframework.runtime.condition.*;
 import org.spockframework.runtime.model.ExpressionInfo;
 import org.spockframework.util.*;
 
 import java.util.*;
 
-import groovy.lang.GString;
-
 public class ExpressionInfoValueRenderer {
-  public static final long MAX_EDIT_DISTANCE_MEMORY = 50 * 1024;
+  private static final List<ExpressionComparisonRenderer> RENDERERS = Arrays.asList(
+    new FailedStringComparisonRenderer(),
+    new FailedSetEqualityComparisonRenderer(),
+    new FailedEqualityComparisonRenderer(),
+    new FailedInstanceOfComparisonRenderer()
+  );
+
   private final ExpressionInfo expr;
 
   private ExpressionInfoValueRenderer(ExpressionInfo expr) {
@@ -86,154 +89,14 @@ public class ExpressionInfoValueRenderer {
   }
 
   private String doRenderValue(ExpressionInfo expr) {
-    String result = renderAsFailedStringComparison(expr);
-    if (result != null) return result;
-
-    result = renderAsFailedSetEqualityComparison(expr);
-    if (result != null) return result;
-
-    result = renderAsFailedEqualityComparison(expr);
-    if (result != null) return result;
-
-
-    result = renderAsFailedInstanceOfComparison(expr);
-    if (result != null) return result;
-
+    for (ExpressionComparisonRenderer renderer : RENDERERS) {
+      String result = renderer.render(expr);
+      if (result != null) return result;
+    }
     return renderToStringOrDump(expr);
   }
 
   private static String renderToStringOrDump(ExpressionInfo expr) {
     return RenderUtil.toStringOrDump(expr.getValue());
-  }
-
-  private String renderAsFailedStringComparison(ExpressionInfo expr) {
-    if (!(Boolean.FALSE.equals(expr.getValue()))) return null;
-    if (!expr.isEqualityComparison(String.class, GString.class)) return null;
-
-    // values can't be null here
-    String str1 = expr.getChildren().get(0).getValue().toString();
-    String str2 = expr.getChildren().get(1).getValue().toString();
-
-    if (((long)str1.length()) * str2.length() > MAX_EDIT_DISTANCE_MEMORY) {
-      return tryReduceStringSizes(str1, str2);
-    } else {
-      return createAndRenderEditDistance(str1, str2);
-    }
-  }
-
-  private String tryReduceStringSizes(String str1, String str2) {
-    int minLength = Math.min(str1.length(), str2.length());
-    int commonStart = minLength;
-    for (int i = 0; i < minLength; i++) {
-      if (str1.charAt(i) != str2.charAt(i)) {
-        commonStart = i-1;
-        break;
-      }
-    }
-    commonStart = Math.max(0, commonStart);
-    int end1 = str1.length()-1;
-    int end2 = str2.length()-1;
-    while (end1 >= 0 && end2 >= 0 && str1.charAt(end1) == str2.charAt(end2)){
-      end1--; end2--;
-    }
-    end1++;
-    end2++;
-
-    if (((long) end1-commonStart) * (end2-commonStart) > MAX_EDIT_DISTANCE_MEMORY) {
-      return "false\nStrings too large to calculate edit distance.";
-    } else {
-      // Check if we can add some context
-      if (((long) end1 - commonStart + 20) * (end2 - commonStart + 20) < MAX_EDIT_DISTANCE_MEMORY){
-        commonStart = Math.max(0, commonStart - 10);
-        end1 = Math.min(str1.length(), end1 + 10);
-        end2 = Math.min(str2.length(), end2 + 10);
-      }
-      return createAndRenderEditDistance(str1, str2, commonStart, end1, end2);
-    }
-  }
-
-  private String createAndRenderEditDistance(String str1, String str2) {
-    EditDistance dist = new EditDistance(str1, str2);
-    return String.format("false\n%d difference%s (%d%% similarity)\n%s",
-      dist.getDistance(), dist.getDistance() == 1 ? "" : "s", dist.getSimilarityInPercent(),
-      new EditPathRenderer().render(str1, str2, dist.calculatePath()));
-  }
-
-  private String createAndRenderEditDistance(String str1, String str2, int commonStart, int end1, int end2) {
-    String sub1 = str1.substring(commonStart, end1);
-    String sub2 = str2.substring(commonStart, end2);
-    EditDistance dist = new EditDistance(sub1, sub2);
-    return String.format("false\n%d difference%s (%d%% similarity) (comparing subset start: %d, end1: %d, end2: %d)\n%s",
-      dist.getDistance(), dist.getDistance() == 1 ? "" : "s", dist.getSimilarityInPercent(),
-      commonStart, end1, end2,
-      new EditPathRenderer().render(sub1, sub2, dist.calculatePath()));
-  }
-
-  @SuppressWarnings("unchecked")
-  private String renderAsFailedSetEqualityComparison(ExpressionInfo expr) {
-    if (!(Boolean.FALSE.equals(expr.getValue()))) return null;
-    if (!expr.isEqualityComparison()) return null;
-
-    Object value1 = expr.getChildren().get(0).getValue();
-    Object value2 = expr.getChildren().get(1).getValue();
-    if (!(value1 instanceof Set) || !(value2 instanceof Set))
-      return null;
-
-    Set v1 = (Set)value1;
-    Set v2 = (Set)value2;
-
-    Set missing = new LinkedHashSet(v2);
-    Set extra = new LinkedHashSet(v1);
-    missing.removeAll(v1);
-    extra.removeAll(v2);
-
-    int missingSize = missing.size();
-    int extraSize = extra.size();
-    int maxDistance = Math.max(v1.size(), v2.size());
-    int editDistance = Math.max(missingSize, extraSize);
-    int similarityInPercent = (maxDistance - editDistance) * 100 / maxDistance;
-    int differences = missingSize + extraSize;
-
-    if (editDistance < 11)
-      return String.format("false%n%d difference%s (%d%% similarity missing %d extra %d)%nmissing: %s%nextra: %s",
-        differences,
-        differences == 1 ? "" : "s",
-        similarityInPercent,
-        missingSize, extraSize,
-        missing, extra);
-
-    return String.format("false%n%d differences (%d%% similarity missing %d extra %d details omitted)",
-      differences,
-      similarityInPercent,
-      missingSize, extraSize);
-  }
-
-  private String renderAsFailedEqualityComparison(ExpressionInfo expr) {
-    if (!(Boolean.FALSE.equals(expr.getValue()))) return null;
-    if (!expr.isEqualityComparison()) return null;
-
-    ExpressionInfo expr1 = expr.getChildren().get(0);
-    ExpressionInfo expr2 = expr.getChildren().get(1);
-    if (expr1.getEffectiveRenderedValue().equals(expr2.getEffectiveRenderedValue())) {
-      addTypeHint(expr1);
-      addTypeHint(expr2);
-    }
-
-    return "false";
-  }
-
-  private String renderAsFailedInstanceOfComparison(ExpressionInfo expr) {
-    if (!(Boolean.FALSE.equals(expr.getValue()))) return null;
-    if (!expr.isInstanceOfComparison()) return null;
-
-    addTypeHint(expr.getChildren().get(0));
-    return "false";
-  }
-
-  private void addTypeHint(ExpressionInfo expr) {
-    if (expr.getRenderedValue() == null) return;
-
-    Class<?> exprType = ObjectUtil.voidAwareGetClass(expr.getValue());
-    expr.setRenderedValue(expr.getRenderedValue() + " (" + exprType.getName() + ")");
   }
 }
